Task 1: changed type of iterator

+++ 0x0E-function_pointers/1-array_iterator.c
@@ -9,7 +9,7 @@
   */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i = -1;
+	size_t i = -1;
 
 	if (array && action)
 		while (++i < size)
